Fix #29: cast numpy images to np.float32

+++ torch_dreams/auto_image_param.py
@@ -1,5 +1,6 @@
 import torch.nn as nn
 import torch
+import numpy as np
 
 from .utils import init_image_param, image_buf_to_rgb
 
@@ -35,5 +36,5 @@ class auto_image_param():
         
     @property
     def rgb(self):
-        rgb = self.to_hwc_tensor().numpy()
+        rgb = self.to_hwc_tensor().numpy().astype(np.float32)
         return rgb
